Move methods in common to mixin module.
